+++ Editor/AppUISettingsProvider.cs
@@ -13,6 +13,28 @@ namespace Unity.AppUI.Editor
         public const string kEditorBuildSettingsConfigKey = "com.unity.dt.app-ui";
         public const string kSettingsPath = "Project/App UI";
 
+        class Styles
+        {
+            public static readonly GUIContent editorOnlyContent = new GUIContent("Editor Only", "Enable this options to prevent the App UI system from running in the player builds.");
+            public static readonly GUIContent autoScaleUIContent = new GUIContent("Auto Scale UI", "Enable this options to correct the scale of UIDocuments, depending on the target platform and screen dpi.");
+            public static readonly GUIContent useCustomEditorUpdateFrequencyContent = new GUIContent("Use Custom Loop Frequency", "Enable this option to override the default update loop frequency (the default frequency is the one used by the Editor loop).");
+            public static readonly GUIContent editorUpdateFrequencyContent = new GUIContent("Update Loop Frequency", "Configure how frequently you want to run the main App UI process loop (to handle queued messages for examples). Default is 60Hz.");
+            public static readonly GUIContent autoOverrideAndroidManifestContent = new GUIContent("Auto Override Android Manifest", "");
+            public static readonly GUIContent enableMacOSGestureRecognitionContent = new GUIContent("Enable Gesture Recognition", "");
+            public static readonly GUIContent includeShadersInPlayerBuildContent = new GUIContent("Include Shaders in Player Build", "");
+
+            public static readonly GUIContent createSettingsAssetContent = new GUIContent("Create Settings Asset", "Create a new App UI settings asset in the project.");
+            public static readonly GUIContent settingsAssetFieldContent = new GUIContent("Settings Asset", "The App UI settings asset to use for this project. If not set, a default one will be created.");
+
+            public static readonly GUIContent editorGroup = new GUIContent("Editor");
+            public static readonly GUIContent runtimeGroup = new GUIContent("Runtime");
+            public static readonly GUIContent androidGroup = new GUIContent("Android");
+            public static readonly GUIContent macOSGroup = new GUIContent("MacOS");
+
+            public static readonly GUIContent androidManifestOverrideWarning = new GUIContent("In order to get all features working properly on Android, you need to override the default Android manifest file with the one provided by App UI.");
+            public static readonly GUIContent macOSGestureRecognitionWarning = new GUIContent("Enabling gesture recognition on MacOS will allow you to use gestures such as pinch, pan, and rotate. However, some gestures are synthesized as mouse events and can't be avoided, such as the Pan gesture.");
+        }
+
         public static void Open()
         {
             SettingsService.OpenProjectSettings(kSettingsPath);
@@ -28,6 +50,7 @@ namespace Unity.AppUI.Editor
             : base(path, scopes)
         {
             label = "App UI";
+            keywords = GetSearchKeywordsFromGUIContentProperties<Styles>();
             s_Instance = this;
         }
 
@@ -36,112 +59,98 @@ namespace Unity.AppUI.Editor
             m_SettingsObject?.Dispose();
         }
 
-        public override void OnTitleBarGUI()
-        {
-            if (EditorGUILayout.DropdownButton(EditorGUIUtility.IconContent("_Popup"), FocusType.Passive, EditorStyles.label))
-            {
-                var menu = new GenericMenu();
-                menu.AddDisabledItem(new GUIContent("Available Settings Assets:"));
-                menu.AddSeparator("");
-                for (var i = 0; i < m_AvailableSettingsAssetsOptions.Length; i++)
-                {
-                    menu.AddItem(new GUIContent(m_AvailableSettingsAssetsOptions[i]),
-                        m_CurrentSelectedAppUISettingsAsset == i, (path) =>
-                        {
-                            Core.AppUI.settings = AssetDatabase.LoadAssetAtPath<AppUISettings>((string) path);
-                        }, m_AvailableAppUISettingsAssets[i]);
-                }
-                menu.AddSeparator("");
-                menu.AddItem(new GUIContent("New Settings Asset…"), false, CreateNewSettingsAsset);
-                menu.ShowAsContext();
-                Event.current.Use();
-            }
-        }
-
         public override void OnGUI(string searchContext)
         {
-            InitializeWithCurrentSettingsIfNecessary();
+            var currentSettingsAsset = m_Settings && EditorUtility.IsPersistent(m_Settings) ? m_Settings : null;
+            var newSettingsAsset = EditorGUILayout.ObjectField(
+                Styles.settingsAssetFieldContent,
+                currentSettingsAsset,
+                typeof(AppUISettings),
+                false) as AppUISettings;
+
+            var removed = !newSettingsAsset && currentSettingsAsset;
+            if (newSettingsAsset != currentSettingsAsset)
+                Core.AppUI.settings = newSettingsAsset ? newSettingsAsset : ScriptableObject.CreateInstance<AppUISettings>();
+
+            InitializeWithCurrentSettingsIfNecessary(removed);
+
+            Debug.Assert(m_Settings);
 
             EditorGUIUtility.labelWidth = 200;
 
-            if (m_AvailableAppUISettingsAssets.Count == 0)
+            if (m_AvailableAppUISettingsAssets.Count == 0 || !EditorUtility.IsPersistent(m_Settings))
             {
+                var notPersistentMessage = EditorUtility.IsPersistent(m_Settings) ? "" : "You have not selected one yet. Please pick one from the field above or create a new one.";
                 EditorGUILayout.HelpBox(
-                    "Settings for App UI are stored in an asset. Click the button below to create a settings asset you can edit.",
+                    $"Settings for App UI are stored in an asset. {notPersistentMessage}\n\nClick the button below to create a settings asset you can edit.",
                     MessageType.Info);
-                if (GUILayout.Button("Create settings asset", GUILayout.Height(30)))
+                if (GUILayout.Button(Styles.createSettingsAssetContent, GUILayout.Height(30)))
                     CreateNewSettingsAsset();
                 GUILayout.Space(20);
             }
 
-            using (new EditorGUI.DisabledScope(m_AvailableAppUISettingsAssets.Count == 0))
+            using (new EditorGUI.DisabledScope(!EditorUtility.IsPersistent(m_Settings)))
             {
                 EditorGUILayout.Space();
                 EditorGUILayout.Separator();
                 EditorGUILayout.Space();
 
-                Debug.Assert(m_Settings != null);
-
                 EditorGUI.BeginChangeCheck();
 
-                EditorGUILayout.LabelField("Editor", EditorStyles.boldLabel);
+                EditorGUILayout.LabelField(Styles.editorGroup, EditorStyles.boldLabel);
 
                 EditorGUI.indentLevel++;
 
-                EditorGUILayout.PropertyField(m_EditorOnly, m_EditorOnlyContent);
+                EditorGUILayout.PropertyField(m_EditorOnly, Styles.editorOnlyContent);
 
-                EditorGUILayout.PropertyField(m_UseCustomEditorUpdateFrequency, m_UseCustomEditorUpdateFrequencyContent);
+                EditorGUILayout.PropertyField(m_UseCustomEditorUpdateFrequency, Styles.useCustomEditorUpdateFrequencyContent);
 
                 using (new EditorGUI.DisabledScope(!m_UseCustomEditorUpdateFrequency.boolValue))
                 {
-                    EditorGUILayout.PropertyField(m_EditorUpdateFrequency, m_EditorUpdateFrequencyContent);
+                    EditorGUILayout.PropertyField(m_EditorUpdateFrequency, Styles.editorUpdateFrequencyContent);
                 }
 
                 EditorGUI.indentLevel--;
 
                 EditorGUILayout.Space();
 
-                EditorGUILayout.LabelField("Runtime", EditorStyles.boldLabel);
+                EditorGUILayout.LabelField(Styles.runtimeGroup, EditorStyles.boldLabel);
 
                 EditorGUI.indentLevel++;
 
-                EditorGUILayout.PropertyField(m_AutoScaleUI, m_AutoScaleUIContent);
+                EditorGUILayout.PropertyField(m_AutoScaleUI, Styles.autoScaleUIContent);
 
                 using (new EditorGUI.DisabledScope(m_EditorOnly.boolValue))
                 {
-                    EditorGUILayout.PropertyField(m_IncludeShadersInPlayerBuild, m_IncludeShadersInPlayerBuildContent);
+                    EditorGUILayout.PropertyField(m_IncludeShadersInPlayerBuild, Styles.includeShadersInPlayerBuildContent);
                 }
 
                 EditorGUI.indentLevel--;
 
                 EditorGUILayout.Space();
 
-                EditorGUILayout.LabelField("Android", EditorStyles.boldLabel);
+                EditorGUILayout.LabelField(Styles.androidGroup, EditorStyles.boldLabel);
 
                 EditorGUI.indentLevel++;
 
-                EditorGUILayout.HelpBox("In order to get all features working properly on Android, " +
-                    "you need to override the default Android manifest file with the one provided by App UI.",
-                    MessageType.Warning);
+                EditorGUILayout.HelpBox(Styles.androidManifestOverrideWarning.text, MessageType.Warning);
 
                 using (new EditorGUI.DisabledScope(m_EditorOnly.boolValue))
                 {
-                    EditorGUILayout.PropertyField(m_AutoOverrideAndroidManifest, m_AutoOverrideAndroidManifestContent);
+                    EditorGUILayout.PropertyField(m_AutoOverrideAndroidManifest, Styles.autoOverrideAndroidManifestContent);
                 }
 
                 EditorGUI.indentLevel--;
 
                 EditorGUILayout.Space();
 
-                EditorGUILayout.LabelField("MacOS", EditorStyles.boldLabel);
+                EditorGUILayout.LabelField(Styles.macOSGroup, EditorStyles.boldLabel);
 
                 EditorGUI.indentLevel++;
 
-                EditorGUILayout.PropertyField(m_EnableMacOSGestureRecognition, m_EnableMacOSGestureRecognitionContent);
+                EditorGUILayout.PropertyField(m_EnableMacOSGestureRecognition, Styles.enableMacOSGestureRecognitionContent);
 
-                EditorGUILayout.HelpBox("Some gestures are synthesized as mouse events on MacOS and can't be avoided, " +
-                    "such as the Pan gesture.",
-                    MessageType.Info);
+                EditorGUILayout.HelpBox(Styles.macOSGestureRecognitionWarning.text, MessageType.Info);
 
                 EditorGUI.indentLevel--;
 
@@ -194,18 +203,18 @@ namespace Unity.AppUI.Editor
             CreateNewSettingsAsset(relativePath);
         }
 
-        void InitializeWithCurrentSettingsIfNecessary()
+        void InitializeWithCurrentSettingsIfNecessary(bool removed)
         {
-            if (Core.AppUI.settings == m_Settings && m_Settings != null && m_SettingsDirtyCount == EditorUtility.GetDirtyCount(m_Settings))
+            if (Core.AppUI.settings == m_Settings && m_Settings && m_SettingsDirtyCount == EditorUtility.GetDirtyCount(m_Settings))
                 return;
 
-            InitializeWithCurrentSettings();
+            InitializeWithCurrentSettings(removed);
         }
 
         /// <summary>
         /// Grab <see cref="AppUISettings"/> and set it up for editing.
         /// </summary>
-        void InitializeWithCurrentSettings()
+        void InitializeWithCurrentSettings(bool removed)
         {
             // Find the set of available assets in the project.
             m_AvailableAppUISettingsAssets = new List<string>(FindAppUISettingsInProject());
@@ -213,70 +222,30 @@ namespace Unity.AppUI.Editor
             // See which is the active one.
             m_Settings = Core.AppUI.settings;
             m_SettingsDirtyCount = EditorUtility.GetDirtyCount(m_Settings);
-            var currentSettingsPath = AssetDatabase.GetAssetPath(m_Settings);
-            if (string.IsNullOrEmpty(currentSettingsPath))
+            if (!EditorUtility.IsPersistent(m_Settings))
             {
-                if (m_AvailableAppUISettingsAssets.Count != 0)
+                if (m_AvailableAppUISettingsAssets.Count != 0 && !removed)
                 {
-                    m_CurrentSelectedAppUISettingsAsset = 0;
                     m_Settings = AssetDatabase.LoadAssetAtPath<AppUISettings>(m_AvailableAppUISettingsAssets[0]);
                     Core.AppUI.settings = m_Settings;
                 }
             }
             else
             {
-                m_CurrentSelectedAppUISettingsAsset = m_AvailableAppUISettingsAssets.IndexOf(currentSettingsPath);
-                if (m_CurrentSelectedAppUISettingsAsset == -1)
-                {
-                    m_AvailableAppUISettingsAssets.Add(currentSettingsPath);
-                    m_CurrentSelectedAppUISettingsAsset = m_AvailableAppUISettingsAssets.IndexOf(currentSettingsPath);
-                }
-
                 ////REVIEW: should we store this by platform?
                 EditorBuildSettings.AddConfigObject(kEditorBuildSettingsConfigKey, m_Settings, true);
-            }
-
-            // Refresh the list of assets we display in the UI.
-            m_AvailableSettingsAssetsOptions = new GUIContent[m_AvailableAppUISettingsAssets.Count];
-            for (var i = 0; i < m_AvailableAppUISettingsAssets.Count; ++i)
-            {
-                var name = m_AvailableAppUISettingsAssets[i];
-                if (name.StartsWith("Assets/"))
-                    name = name.Substring("Assets/".Length);
-                if (name.EndsWith(".asset"))
-                    name = name.Substring(0, name.Length - ".asset".Length);
-
-                // Ugly hack: GenericMenu interprets "/" as a submenu path. But luckily, "/" is not the only slash we have in Unicode.
-                m_AvailableSettingsAssetsOptions[i] = new GUIContent(name.Replace("/", "\u29f8"));
             }
 
             // Look up properties.
             m_SettingsObject = new SerializedObject(m_Settings);
 
             m_EditorOnly = m_SettingsObject.FindProperty("m_EditorOnly");
-            m_EditorOnlyContent = new GUIContent("Editor Only",
-                "Enable this options to prevent the App UI system from running in the player builds.");
-
             m_AutoScaleUI = m_SettingsObject.FindProperty("m_AutoCorrectUiScale");
-            m_AutoScaleUIContent = new GUIContent("Auto Scale UI",
-                "Enable this options to correct the scale of UIDocuments, depending on the target platform and screen dpi.");
-
             m_UseCustomEditorUpdateFrequency = m_SettingsObject.FindProperty("m_UseCustomEditorUpdateFrequency");
-            m_UseCustomEditorUpdateFrequencyContent = new GUIContent("Use Custom Loop Frequency",
-                "Enable this option to override the default update loop frequency (the default frequency is the one used by the Editor loop).");
-
             m_EditorUpdateFrequency = m_SettingsObject.FindProperty("m_EditorUpdateFrequency");
-            m_EditorUpdateFrequencyContent = new GUIContent("Update Loop Frequency",
-                "Configure how frequently you want to run the main App UI process loop (to handle queued messages for examples). Default is 60Hz.");
-
             m_AutoOverrideAndroidManifest = m_SettingsObject.FindProperty("m_AutoOverrideAndroidManifest");
-            m_AutoOverrideAndroidManifestContent = new GUIContent("Auto Override Android Manifest", "");
-
             m_EnableMacOSGestureRecognition = m_SettingsObject.FindProperty("m_EnableMacOSGestureRecognition");
-            m_EnableMacOSGestureRecognitionContent = new GUIContent("Enable Gesture Recognition", "");
-
             m_IncludeShadersInPlayerBuild = m_SettingsObject.FindProperty("m_IncludeShadersInPlayerBuild");
-            m_IncludeShadersInPlayerBuildContent = new GUIContent("Include Shaders in Player Build", "");
         }
 
         void Apply()
@@ -284,7 +253,7 @@ namespace Unity.AppUI.Editor
             if (!m_Settings)
                 return;
 
-            m_SettingsObject.ApplyModifiedProperties();
+            m_SettingsObject.ApplyModifiedPropertiesWithoutUndo();
             m_SettingsObject.Update();
             m_Settings.OnChange();
         }
@@ -309,7 +278,6 @@ namespace Unity.AppUI.Editor
         }
 
         [SerializeField] AppUISettings m_Settings;
-        [SerializeField] bool m_SettingsIsNotAnAsset;
 
         [NonSerialized] int m_SettingsDirtyCount;
         [NonSerialized] SerializedObject m_SettingsObject;
@@ -322,18 +290,6 @@ namespace Unity.AppUI.Editor
         [NonSerialized] SerializedProperty m_IncludeShadersInPlayerBuild;
 
         [NonSerialized] List<string> m_AvailableAppUISettingsAssets;
-        [NonSerialized] GUIContent[] m_AvailableSettingsAssetsOptions;
-        [NonSerialized] int m_CurrentSelectedAppUISettingsAsset;
-
-        [NonSerialized] GUIStyle m_NewAssetButtonStyle;
-
-        GUIContent m_EditorOnlyContent;
-        GUIContent m_AutoScaleUIContent;
-        GUIContent m_UseCustomEditorUpdateFrequencyContent;
-        GUIContent m_EditorUpdateFrequencyContent;
-        GUIContent m_AutoOverrideAndroidManifestContent;
-        GUIContent m_EnableMacOSGestureRecognitionContent;
-        GUIContent m_IncludeShadersInPlayerBuildContent;
 
         static AppUISettingsProvider s_Instance;