+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-pre.20] - 2025-06-10
+
+### Removed
+
+- Removed `Canvas.scrollDirection` property in order to support correctly the scroll direction set in the Operating System's preferences.
+
+### Fixed
+
+- Fixed tooltip that weren't showing up anymore since last pre-release.
+- Fixed `VisualElementExtensions.IsInvisible` method to check recursively in ancestors if any element is considered as _invisible_.
+
+### Changed
+
+- Use `resolvedStyle.translate` instead of `transform.position` to move elements such as the `Canvas` container.
+- Replaced the dropdown with an ObjectField to pick an available App UI settings asset.
+
+### Added
+
+- Added a safety check in TooltipManipulator to ensure the anchor element of the scheduled tooltip is still valid (not _invisible_) when it is time to _animate in_ the tooltip.
+- Added monitoring of key press events in the TooltipManipulator in order to dismiss any existing tooltip when the user interacts with the keyboard.
+
 ## [2.0.0-pre.19] - 2025-04-22
 
 ### Fixed